move signal handler for task status change
