+++ qhana_plugin_runner/tasks.py
@@ -1,10 +1,8 @@
-import urllib.parse
 from datetime import datetime
 from blinker import Namespace
 
 from celery.result import AsyncResult
 from celery.utils.log import get_task_logger
-import requests
 
 from qhana_plugin_runner.db.models.tasks import ProcessingTask
 
@@ -16,27 +14,7 @@ TASK_LOGGER = get_task_logger(_name)
 
 TASK_SIGNALS = Namespace()
 
-
-def task_status_changed_handler(sender, db_id: int):
-    """Emit signal to all registered callback urls that the status of a task has changed.
-    Attributes:
-        sender: The sender of the signal.
-        db_id (int): The database id of the task.
-    """
-    task_data: ProcessingTask = ProcessingTask.get_by_id(id_=db_id)
-    if task_data is None:
-        raise KeyError(
-            f"Could not find db entry for id {db_id}, failed to emit signal on status change!"
-        )
-
-    callback_urls = task_data.data.get("status_changed_callback_urls", [])
-    task_url = task_data.data.get("task_view", None)
-    for callback_url in callback_urls:
-        requests.post(callback_url, json={"url": task_url, "status": task_data.status})
-
-
 TASK_STATUS_CHANGED = TASK_SIGNALS.signal("task-status-changed")
-TASK_STATUS_CHANGED.connect(task_status_changed_handler)
 
 
 # TODO add periodic cleanup task to remove old results from the database!
